+++ Changes.md
@@ -14,6 +14,12 @@ Improvements
 
 - FrameMask : Improved performance when dealing with long frame ranges.
 - Node : Improved performance of type queries related to dispatch processes.
+- FilterResults : Improved performance of hash.
+
+API
+---
+
+- SceneAlgo : Added matchingPathsHash.
 
 0.59.3.0 (relative to 0.59.2.0)
 ========